+++ packages/runtime/src/runtime/hooks/useNavigteExternalScreen.ts
@@ -1,18 +1,25 @@
 import { useMemo } from "react";
-import { type NavigateExternalScreen } from "@ensembleui/react-framework";
+import {
+  useEvaluate,
+  type NavigateExternalScreen,
+} from "@ensembleui/react-framework";
+import { isString } from "lodash-es";
 import { openExternalScreen } from "../navigation";
 import { type EnsembleActionHook } from "./useEnsembleAction";
 
 export const useNavigateExternalScreen: EnsembleActionHook<
   NavigateExternalScreen
 > = (action) => {
+  const evaluatedAction = useEvaluate(
+    isString(action) ? { url: action } : { ...action },
+  );
   const callback = useMemo(() => {
-    if (!action) {
+    if (!evaluatedAction || !evaluatedAction.url) {
       return;
     }
 
-    return () => openExternalScreen(action);
-  }, [action]);
+    return () => openExternalScreen(evaluatedAction as NavigateExternalScreen);
+  }, [evaluatedAction]);
 
   return callback ? { callback } : undefined;
 };