+++ apps/kitchen-sink/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @ensembleui/react-kitchen-sink
 
+## 0.0.26
+
+### Patch Changes
+
+- 0e65e56: add onChange action in Checkbox widget
+- Updated dependencies [0e65e56]
+- Updated dependencies [392d9d1]
+  - @ensembleui/react-framework@0.1.26
+  - @ensembleui/react-runtime@0.1.35
+
+## 0.0.25
+
+### Patch Changes
+
+- Updated dependencies [de7d06b]
+  - @ensembleui/react-framework@0.1.25
+  - @ensembleui/react-runtime@0.1.34
+
+## 0.0.24
+
+### Patch Changes
+
+- f44932f: Added support for multiple theme support
+- 6b4b2e8: added support to return the API response in ensemble.invokeAPI
+- Updated dependencies [f44932f]
+- Updated dependencies [6b4b2e8]
+  - @ensembleui/react-framework@0.1.24
+  - @ensembleui/react-runtime@0.1.33
+
 ## 0.0.23
 
 ### Patch Changes