+++ apps/starter/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ensembleui/react-starter
 
+## 0.3.24
+
+### Patch Changes
+
+- Updated dependencies [0e65e56]
+- Updated dependencies [392d9d1]
+  - @ensembleui/react-framework@0.1.26
+  - @ensembleui/react-runtime@0.1.35
+
+## 0.3.23
+
+### Patch Changes
+
+- Updated dependencies [de7d06b]
+  - @ensembleui/react-framework@0.1.25
+  - @ensembleui/react-runtime@0.1.34
+
+## 0.3.22
+
+### Patch Changes
+
+- Updated dependencies [f44932f]
+- Updated dependencies [6b4b2e8]
+  - @ensembleui/react-framework@0.1.24
+  - @ensembleui/react-runtime@0.1.33
+
 ## 0.3.21
 
 ### Patch Changes